Listener interface updated

+++ src/creator.py
@@ -60,7 +60,7 @@ while interface.running:
 
             csv_file_name = file_names.get_csv_file_path()
             wav_file_name = file_names.get_wav_file_path()
-            if os.path.isfile(csv_file_name):
+            if os.path.isfile(csv_file_name) and os.path.isfile(csv_file_name):
                 file_names.increase_number_of_recordings_created()
                 sf.write(wav_file_name, recording.get_data(), audio_io.sampling_freq)
                 pd.DataFrame(positional_data).to_csv(csv_file_name, header=None, index=None)
